requirements.txt
sentence-transformers
transformers
torch
accelerate
pinecone
fastapi
uvicorn
python-dotenv